[RateLimiter] Moved classes implementing LimiterInterface to a new namespace
