+++ src/Symfony/Component/RateLimiter/Policy/Window.php
@@ -9,7 +9,9 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\Component\RateLimiter;
+namespace Symfony\Component\RateLimiter\Policy;
+
+use Symfony\Component\RateLimiter\LimiterStateInterface;
 
 /**
  * @author Wouter de Jong <wouter@wouterj.nl>
